+++ modules/admin/storage/Image.php
@@ -2,39 +2,50 @@
 
 namespace admin\storage;
 
-use \admin\models\StorageImage;
+use Yii;
+use Exception;
+use \Imagine\Gd\Imagine;
+use admin\models\StorageImage;
+use admin\models\StorageFilter;
+use luya\helpers\ArrayHelper;
 
 class Image
 {
     public function create($fileId, $filterId = 0)
     {
+        // find image for file_id and filter_id
         $img = StorageImage::find()->where(['file_id' => $fileId, 'filter_id' => $filterId])->asArray()->one();
-        if($img) {
+        
+        // the image exists already, we return the already created image id
+        if ($img) {
             return $img['id'];
-        }   
-        $file = \yii::$app->storage->file->getPath($fileId);
-        $info = \yii::$app->storage->file->getInfo($fileId);
+        }
+           
+        $file = Yii::$app->storage->file->getPath($fileId);
+        $info = Yii::$app->storage->file->getInfo($fileId);
+        
         try {
-            $imagine = new \Imagine\Gd\Imagine();
+            $imagine = new Imagine();
             $image = $imagine->open($file);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return $e->getMessage();
         }
+        
         $fileName = $filterId.'_'.$info->name_new_compound;
 
         if (empty($filterId)) {
-            $save = $image->save(\yii::$app->storage->dir.$fileName);
+            $save = $image->save(Yii::$app->storage->dir.$fileName);
         } else {
-            $model = \admin\models\StorageFilter::find()->where(['id' => $filterId])->one();
+            $model = StorageFilter::find()->where(['id' => $filterId])->one();
             if (!$model) {
-                throw new \Exception("could not find the provided filter id '$filterId'.");
+                throw new Exception("Could not find the provided filter id '$filterId'.");
             }
             $newimage = $model->applyFilter($image, $imagine);
-            $save = $newimage->save(\yii::$app->storage->dir.$fileName);
+            $save = $newimage->save(Yii::$app->storage->dir.$fileName);
         }
 
         if ($save) {
-            $model = new \admin\models\StorageImage();
+            $model = new StorageImage();
             $model->setAttributes([
                 'file_id' => $fileId,
                 'filter_id' => $filterId,
@@ -57,7 +68,7 @@ class Image
     public function filterApply($imageId, $filterIdentifier, $returnImageObject = false)
     {
         // resolve $filterIdentifier
-        $filter = \admin\models\StorageFilter::find()->where(['identifier' => $filterIdentifier])->asArray()->one();
+        $filter = StorageFilter::find()->where(['identifier' => $filterIdentifier])->asArray()->one();
         if (!$filter) {
             throw new \Exception('could not find the filterIdentifier '.$filterIdentifier);
         }
@@ -83,18 +94,23 @@ class Image
         
         return $imageObject->source;
     }
+    
+    public function getSource($imageId)
+    {
+        return ($image = $this->get($imageId)) ? $image->source : '';
+    }
 
     // @web/storage/the-originame_name_$filterId_$fileIdf.jpg
     public function get($imageId)
     {
         // get the real full image path to display this file.
-        $data = \admin\models\StorageImage::find()->where(['id' => $imageId])->with('file')->one();
+        $data = StorageImage::find()->where(['id' => $imageId])->with('file')->one();
         if (!$data) {
             return false;
         }
         $fileName = implode([$data->filter_id, $data->file->name_new_compound], '_');
 
-        return \luya\helpers\ArrayHelper::toObject([
+        return ArrayHelper::toObject([
             'filter_id' => $data->filter_id,
             'file_id' => $data->file_id,
             'image_id' => $data->id,